% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_boundary_by_area_id}
\alias{get_boundary_by_area_id}
\title{Get boundary by area ID}
\usage{
get_boundary_by_area_id(client, area_id)
}
\arguments{
\item{client}{A client from \code{client()} authenticated with \code{sign_in()}.}

\item{area_id}{UUID area identifier.}
}
\value{
A tibble with boundary geometry.
}
\description{
Get boundary by area ID
}
