% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_boundary_by_location_id}
\alias{get_boundary_by_location_id}
\title{Get boundary by location ID}
\usage{
get_boundary_by_location_id(client, location_id)
}
\arguments{
\item{client}{A client from \code{client()} authenticated with \code{sign_in()}.}

\item{location_id}{Location identifier.}
}
\value{
A tibble with boundary geometry.
}
\description{
Get boundary by location ID
}
