% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_boundary_by_osm_id}
\alias{get_boundary_by_osm_id}
\title{Get boundary by OpenStreetMap ID}
\usage{
get_boundary_by_osm_id(client, osm_id)
}
\arguments{
\item{client}{A client from \code{client()} authenticated with \code{sign_in()}.}

\item{osm_id}{Numeric OSM identifier.}
}
\value{
A tibble with boundary geometry.
}
\description{
Get boundary by OpenStreetMap ID
}
