% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_remaining_queries}
\alias{get_remaining_queries}
\title{Get remaining monthly query quota}
\usage{
get_remaining_queries(client)
}
\arguments{
\item{client}{A client from \code{client()} authenticated with \code{sign_in()}.}
}
\value{
A tibble with remaining quota.
}
\description{
Get remaining monthly query quota
}
