% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_request_status}
\alias{get_request_status}
\title{Get status of a submitted request}
\usage{
get_request_status(client, request_id)
}
\arguments{
\item{client}{A client from \code{client()} authenticated with \code{sign_in()}.}

\item{request_id}{Request UUID.}
}
\value{
A tibble with one row.
}
\description{
Get status of a submitted request
}
