% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_subscription}
\alias{get_subscription}
\title{Get current subscription profile}
\usage{
get_subscription(client)
}
\arguments{
\item{client}{A client from \code{client()} authenticated with \code{sign_in()}.}
}
\value{
A tibble with profile row.
}
\description{
Get current subscription profile
}
