% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{si_auth_signin}
\alias{si_auth_signin}
\title{Sign in to Supabase with email/password}
\usage{
si_auth_signin(client, email, password)
}
\arguments{
\item{client}{A client from \code{si_client()}.}

\item{email}{User email.}

\item{password}{User password.}
}
\value{
A list with authenticated \code{client}, \code{session}, and raw \code{data}.
}
\description{
Sign in to Supabase with email/password
}
