% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{si_get_polygon_by_location_id}
\alias{si_get_polygon_by_location_id}
\title{Get one bounds polygon by location ID (via secure RPC)}
\usage{
si_get_polygon_by_location_id(client, location_id)
}
\arguments{
\item{client}{A client from \code{si_client()} authenticated with \code{si_auth_signin()}.}

\item{location_id}{Text location identifier from \code{public.location.location_id}.}
}
\value{
A tibble with up to one row.
}
\description{
Get one bounds polygon by location ID (via secure RPC)
}
