% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{si_get_polygon_by_osm_id}
\alias{si_get_polygon_by_osm_id}
\title{Get one bounds polygon by OSM ID}
\usage{
si_get_polygon_by_osm_id(client, osm_id)
}
\arguments{
\item{client}{A client from \code{si_client()} authenticated with \code{si_auth_signin()}.}

\item{osm_id}{Numeric OSM identifier in \code{public.bounds.osm_id}.}
}
\value{
A tibble with up to one row.
}
\description{
Get one bounds polygon by OSM ID
}
