% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{sign_in}
\alias{sign_in}
\title{Sign in with email and password}
\usage{
sign_in(client, email, password)
}
\arguments{
\item{client}{A client from \code{client()}.}

\item{email}{User email.}

\item{password}{User password.}
}
\value{
A list with authenticated \code{client}, \code{session}, and raw \code{data}.
}
\description{
Sign in with email and password
}
