% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_themes}
\alias{get_themes}
\title{List available social-infrastructure themes}
\usage{
get_themes(client)
}
\arguments{
\item{client}{A client from \code{client()}.}
}
\value{
A tibble with theme IDs and types.
}
\description{
List available social-infrastructure themes
}
